promptsail/pyproject.toml
[tool.poetry]
name = "promptsail-explore"
version = "0.1.0"
description = ""
authors = ["Sergey Panfilov <sergray@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.41.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
